Refactor block registration and move Minecraft block handling.
